授课成果查询-后台

+++ src/main/java/com/beifen/edu/administration/dao/Edu8001Dao.java
@@ -21,14 +21,14 @@ public interface Edu8001Dao extends JpaRepository<Edu8001, Long>, JpaSpecificati
     void deleteFinanceInfodetail(List<String> edu108IdList);
 
     @Query(value = "select case when sum(fy) is null then 0 else sum(fy) end from EDU8001 where SUBSTR(pay_time, 1, 4) = ?1",nativeQuery = true)
-    Integer findByYear(String year);
+    Double findByYear(String year);
 
     @Query(value = "select case when sum(fy) is null then 0 else sum(fy) end from EDU8001 where lbbm = ?1",nativeQuery = true)
-    Integer findMoneyByLx(String ejdm);
+    Double findMoneyByLx(String ejdm);
 
     @Query(value = "select count(*) from EDU8001 where lbbm = ?1",nativeQuery = true)
     Integer findCountByLx(String ejdm);
 
     @Query(value = "select case when sum(fy) is null then 0 else sum(fy) end from EDU8001 ",nativeQuery = true)
-    Integer findMoney();
+    Double findMoney();
 }
